First stab at a checkSpam function
Fail with an error if spamVerdict or virusVerdict are not "PASS".
This might be a bit too extreme, but some spam messages have
definitely crept through as "PASS" so not sure that any lower
setting makes sense.

Failing with error isn't particularly clean, but is easy.  Future
work could try to add a send/don't send decision and skip the
sendMessage step.